requirements.txt
numpy==1.14.1
scipy==1.0.0
flask==0.12.2
flask_cors==3.0.3
progressbar2==3.37.0
